+++ modules/sdk/gradle-plugins-defaults/CHANGELOG.markdown
@@ -316,8 +316,27 @@ include the projects containing a `.lfrbuild-portal-pre` marker file.
 ### Changed
 - [LPS-67352]: Update the [Liferay Gradle Plugins] dependency to version 2.0.45.
 
+## 1.2.36 - 2016-09-29
+
+### Changed
+- [LPS-58672]: Update the [Liferay Gradle Plugins] dependency to version 2.0.46.
+
+## 1.2.37 - 2016-09-30
+
+### Changed
+- [LPS-67352]: Update the [Liferay Gradle Plugins] dependency to version 2.0.47.
+
+## 1.2.38 - 2016-09-30
+
+### Added
+- [LPS-68448]: Fail the build of an OSGi project if the version in the
+`npm-shrinkwrap.json` file does not match the project version.
+- [LPS-68448]: The task `updateVersion` of OSGi and theme projects updates the
+version in the `npm-shrinkwrap.json` file, if present.
+
 [Liferay Gradle Plugins]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins
 [Liferay Gradle Plugins Node]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-node
+[LPS-58672]: https://issues.liferay.com/browse/LPS-58672
 [LPS-61099]: https://issues.liferay.com/browse/LPS-61099
 [LPS-66853]: https://issues.liferay.com/browse/LPS-66853
 [LPS-66906]: https://issues.liferay.com/browse/LPS-66906
@@ -337,5 +356,6 @@ include the projects containing a `.lfrbuild-portal-pre` marker file.
 [LPS-68297]: https://issues.liferay.com/browse/LPS-68297
 [LPS-68305]: https://issues.liferay.com/browse/LPS-68305
 [LPS-68306]: https://issues.liferay.com/browse/LPS-68306
+[LPS-68448]: https://issues.liferay.com/browse/LPS-68448
 [LRDOCS-2841]: https://issues.liferay.com/browse/LRDOCS-2841
 [LRDOCS-2981]: https://issues.liferay.com/browse/LRDOCS-2981